+++ IsekaiMod/Content/Heritages/IsekaiAngel/IsekaiAngelHeritage.cs
@@ -8,11 +8,12 @@ using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
 using Kingmaker.Enums.Damage;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
+using Kingmaker.UnitLogic.ActivatableAbilities;
 using Kingmaker.UnitLogic.FactLogic;
 using Kingmaker.UnitLogic.Mechanics;
 using Kingmaker.UnitLogic.Mechanics.Components;
 
-namespace IsekaiMod.Content.Heritages
+namespace IsekaiMod.Content.Heritages.IsekaiAngel
 {
     class IsekaiAngelHeritage
     {
@@ -20,6 +21,8 @@ namespace IsekaiMod.Content.Heritages
         {
             var AngelicBoltAbility = Resources.GetModBlueprint<BlueprintAbility>("AngelicBoltAbility");
             var AngelWingsFeature = Resources.GetBlueprint<BlueprintFeature>("d9bd0fde6deb2e44a93268f2dfb3e169");
+            var BlackWingsAbility = Resources.GetModBlueprint<BlueprintActivatableAbility>("BlackWingsAbility");
+            var GhostWingsAbility = Resources.GetModBlueprint<BlueprintActivatableAbility>("GhostWingsAbility");
 
             // Angel Heritage
             var Icon_Angel = AssetLoader.LoadInternal("Heritages", "ICON_ANGEL.png");
@@ -119,7 +122,9 @@ namespace IsekaiMod.Content.Heritages
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] {
                         AngelicBoltAbility.ToReference<BlueprintUnitFactReference>(),
-                        AngelWingsFeature.ToReference<BlueprintUnitFactReference>()
+                        AngelWingsFeature.ToReference<BlueprintUnitFactReference>(),
+                        BlackWingsAbility.ToReference<BlueprintUnitFactReference>(),
+                        GhostWingsAbility.ToReference<BlueprintUnitFactReference>(),
                     };
                 });